Move MasterFailureTest to src/test, and remove redundant TestOutputStream
